GitHub Link Dialog moved to components in sections
